task/constants.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

DEFAULT_SYSTEM_PROMPT = "You are an assistant who answers concisely and informatively."
DIAL_ENDPOINT = "https://ai-proxy.lab.epam.com"
API_KEY = os.getenv('MY_DIAL_API_KEY', '')
