requirements.txt
requests==2.28.0
aiohttp==3.13.2
aidial-client==0.3.0
python-dotenv==1.1.0
